Use doubles instead of floats for streamflow quantities. Flow.cpp, .h - Change types in calling sequences for SetGeometry(), GetDepthFromQ(), GetExtraSvFluxValue(), GetDischarge(), GetUpstreamFlow(). Declare many member variables and local variables as doubles instead of floats. FluxExpr.cpp - Change one local variables from float to double. GlobalMethods.h -  Change types in calling sequences for GetLateralInflow(), GetReachFluxes(), GetReachInflow(), GetReachOutflow(), GetLateralSVInflow(), GetReachSVOutflow(), GetReachSVInflow(). ReachRouting.cpp - Change some locals from floats to doubles. Change types in calling sequences for GetLateralInflow(), GetReachFluxes(), EstimateReachOutflow(), GetReachInflow(), GetReachOutflow(), GetLateralSVInflow(), GetReachSVOutflow(). WaterRights.cpp, .h - Change calling sequence for GetAvailableSourceFlow(). Change some locals from floats to doubles.

+++ SourceCode/Flow/FluxExpr.cpp
@@ -480,7 +480,7 @@ bool FluxExpr::Step( FlowContext *pFlowContext )
             Reach *pReach = pModel->GetReachFromStreamIndex( sourceIndex );
             ASSERT( pReach != NULL );
 
-            float availSourceFlow = pReach->m_availableDischarge;  // m3/sec
+            double availSourceFlow = pReach->m_availableDischarge;  // m3/sec
 
             // make sure min flows are maintained
             availSourceFlow -= minFlow;  // m3/sec
@@ -566,7 +566,7 @@ bool FluxExpr::Step( FlowContext *pFlowContext )
 
                if ( totalSinkDemand > availSourceFlow * m_withdrawalCutoffCoeff )   // 0.1 - to take care of low flow problem with stream routing
                   {
-                  totalSatisfiedDemand   = availSourceFlow * m_withdrawalCutoffCoeff;
+                  totalSatisfiedDemand   = (float)(availSourceFlow * m_withdrawalCutoffCoeff);
                   totalUnsatisfiedDemand = totalSinkDemand-totalSatisfiedDemand;
                   }
