+++ SourceCode/Flow/GlobalMethods.h
@@ -199,22 +199,22 @@ protected:
    bool  SolveReachKinematicWave( FlowContext* );
    bool  SolveReach2KW(FlowContext*);
 
-   float GetLateralInflow( Reach *pReach );
+   double GetLateralInflow( Reach *pReach );
    WaterParcel GetLateralInflowWP(Reach* pReach);
-   float GetReachFluxes( FlowContext*, Reach *pReach );
+   double GetReachFluxes( FlowContext*, Reach *pReach );
 
-   float GetReachInflow( Reach *pReach, int subNode );
+   double GetReachInflow( Reach *pReach, int subNode );
    WaterParcel GetReachInflowWP(Reach * pReach, int subnode);
-   float GetReachOutflow(ReachNode* pReachNode);
+   double GetReachOutflow(ReachNode* pReachNode);
    WaterParcel GetReachOutflowWP(ReachNode* pReachNode);
 
-   float EstimateReachOutflow( Reach *pReach, int i, double timeStep, float lateralInflow);
+   double EstimateReachOutflow( Reach *pReach, int i, double timeStep, double lateralInflow);
    WaterParcel ApplyReachOutflowWP(Reach *pReach, int subnode, double timeStep, WaterParcel lateralInflowWP); 
    double KinematicWave(double oldQ_cms, double upstreamInflow_cms, double lateralInflow_cms, Reach * pReach);
 
-   float GetLateralSVInflow( Reach *pReach, int sv );
-   float GetReachSVOutflow( ReachNode *pReachNode, int sv );   // recursive!!! for pahntom nodes
-   float GetReachSVInflow( Reach *pReach, int subNode, int sv );
+   double GetLateralSVInflow( Reach *pReach, int sv );
+   double GetReachSVOutflow( ReachNode *pReachNode, int sv );   // recursive!!! for pahntom nodes
+   double GetReachSVInflow( Reach *pReach, int subNode, int sv );
 
    static void GetReachDerivatives( double time, double timestep, int svCount, double *derivatives, void *extra );
